// app/lib/types.ts
export interface Circuit {
  id: string;
  description: string;
  load_type: string;
  rated_power_kw: string;
  voltage: string;
  phases: string;
  flc: string;
  breaker_size: string;
  cable_size: string;
  cable_type: string;
  starter_type: string;
  protection: string;
  bms: string;
  duty: string;
  notes: string;
}